Update point-of-sale.md
Update the Testing your integration section with a reference to the general guidelines that have been updated

Bundled up the additional external articles at the bottom of this section. I do think that they should be somewhere else, but not sure what the best process is yet and we don't have the right data yet to be able to point integration partners to the right resources. So please approve this for the time being and we will be looking for a better solution before the end of the quarter