Add CreatorProfile & UpdaterProfile explanation in orderitems.md (#435)
* Update orderitems.md

Some endpoints include now "Creator Profile" and "Updater Profile" in the response. This is already displayed in the code example in the documentation, but not explained. I noticed it also in bills/getAll.
Can we please add an explanation to all endpoints that return those values?

* Update orderitems.md

Edited Britta's entry for `CreatorProfileId` and `UpdaterProfileId`, and added missing definition for top level response object.

* Update README.md

Added Changelog entry

---------

Co-authored-by: Mike Adams <92085514+MikeAdamsMews@users.noreply.github.com>